JWWEB-737 Renamed from TransactionDetails to HistoryItemDetails
