+++ src/components/HistoryItemDetails/components/AssetItemPreview/assetItemPreview.m.scss
