+++ src/components/HistoryItemDetails/components/FieldPreview/fieldPreview.m.scss
