+++ src/components/HistoryItemDetails/historyItemDetails.m.scss
