+++ src/components/HistoryItemDetails/templates/index.js
