+++ src/components/HistoryItemDetails/utils/handleEditNote.js
