+++ src/pages/HistoryItem/HistoryItemView.js
@@ -2,7 +2,7 @@
 
 import React from 'react'
 import { t } from 'ttag'
-import { TransactionDetails } from 'components/TransactionDetails/TransactionDetails'
+import { HistoryItemDetails } from 'components/HistoryItemDetails/HistoryItemDetails'
 import { JLink } from 'components/base'
 
 import offset from 'styles/offsets.m.scss'
@@ -16,7 +16,7 @@ type Props = {
 export function HistoryItemView(props: Props) {
   return (
     <div className={style.core}>
-      <TransactionDetails className={offset.mb32} txHash={props.itemId} />
+      <HistoryItemDetails className={offset.mb32} txHash={props.itemId} />
       <JLink
         theme='button-general'
         href='/history'
